feat: improve role test authentication and fix test suite
- Extract reusable Auth0 authentication functions to test/helpers/auth.ts
- Refactor role-update.spec.ts to use shared authentication helpers
- Fix role.spec.ts token acquisition using proper Auth0 flow
- Remove test skipping from 200/201 integration tests
- Implement proper authentication failure handling with graceful skips
- Add comprehensive Role Management API documentation to README.md
- Improve error message handling with additional expected error types
- Enable all role tests to run without .skip directives
